fix: proper support for buffer overflow
When the buffer overflow we stop the subscription and wait for at least 10% of
the buffer to be available before restarting the subscriptions. Before we were
throwing an exception and failing the stage, the issue was that we had stalled
graphs which were never restarted since we don't implement the supervision
strategy properly.